Update README.md
I could not get the Video component to work by following the provided example. I kept getting an error message saying "React.createElement: type should not be null, undefined, boolean, or number.". When I changed the way I include the component, it started working. It might be related to recent versions of react-native, not sure. You're welcome to modify the suggested text if you like. Thanks for a great piece of work!